Enhance file transfer functionality and add exit request handling
- Update CMake configuration to support code coverage for tests.
- Introduce DMLOG_FLAG_EXIT_REQUESTED to signal monitor exit.
- Implement dmlog_exit_monitor function to handle exit requests.
- Modify file transfer structures to include status and improve error handling.
- Update monitor to process file transfer requests and handle errors more effectively.
- Refactor memory read/write functions to use 64-bit addresses for compatibility.

+++ tools/monitor/backend.c
@@ -48,7 +48,7 @@ int backend_disconnect(backend_type_t type, int socket)
     return backends[type]->disconnect(socket);
 }
 
-int backend_read_memory(backend_type_t type, int socket, uint32_t address, void *buffer, size_t length)
+int backend_read_memory(backend_type_t type, int socket, uint64_t address, void *buffer, size_t length)
 {
     if(type >= BACKEND_TYPE__COUNT || backends[type] == NULL)
     {
@@ -57,7 +57,7 @@ int backend_read_memory(backend_type_t type, int socket, uint32_t address, void 
     return backends[type]->read_memory(socket, address, buffer, length);
 }
 
-int backend_write_memory(backend_type_t type, int socket, uint32_t address, const void *buffer, size_t length)
+int backend_write_memory(backend_type_t type, int socket, uint64_t address, const void *buffer, size_t length)
 {
     if(type >= BACKEND_TYPE__COUNT || backends[type] == NULL)
     {
