+++ tools/monitor/gdb.h
@@ -14,7 +14,7 @@ extern int gdb_connect(const backend_addr_t *addr);
 extern int gdb_disconnect(int socket);
 extern int gdb_continue(int socket);
 extern int gdb_resume_briefly(int socket);
-extern int gdb_read_memory(int socket, uint32_t address, void *buffer, size_t length);
-extern int gdb_write_memory(int socket, uint32_t address, const void *buffer, size_t length);
+extern int gdb_read_memory(int socket, uint64_t address, void *buffer, size_t length);
+extern int gdb_write_memory(int socket, uint64_t address, const void *buffer, size_t length);
 
 #endif // GDBSERVER_H
